requirements.txt
gmusicapi
python-vlc
pysimplegui